edit/normal item component
